backend/slm-service/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM Libraries - Updated for Qwen2 support
llama-cpp-python==0.2.77
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
accelerate==0.25.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
numpy==1.24.3
pandas==2.1.3